+++ src/main/java/com/plivo/api/models/base/Creator.java
@@ -2,6 +2,7 @@ package com.plivo.api.models.base;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.plivo.api.PlivoClient;
 import com.plivo.api.exceptions.PlivoRestException;
 import java.io.IOException;
 import retrofit2.Call;
@@ -26,6 +27,13 @@ public abstract class Creator<CreateResponse extends BaseResponse> extends BaseR
 
     return response.body();
   }
+
+  @Override
+  public Creator<CreateResponse> client(final PlivoClient plivoClient) {
+    this.plivoClient = plivoClient;
+    return this;
+  }
+
 
   protected abstract Call<CreateResponse> obtainCall();
 }