+++ src/main/java/com/plivo/api/models/call/actions/CallRecordCreator.java
@@ -1,5 +1,6 @@
 package com.plivo.api.models.call.actions;
 
+import com.plivo.api.PlivoClient;
 import com.plivo.api.exceptions.PlivoRestException;
 import com.plivo.api.models.base.Creator;
 import java.io.IOException;
@@ -20,4 +21,11 @@ public class CallRecordCreator extends Creator<CallRecordCreateResponse> {
   public CallRecordCreateResponse record() throws IOException, PlivoRestException {
     return create();
   }
+
+  @Override
+  public CallRecordCreator client(final PlivoClient plivoClient) {
+    this.plivoClient = plivoClient;
+    return this;
+  }
+
 }