+++ src/main/java/com/plivo/api/models/conference/ConferenceMemberSpeakDeleter.java
@@ -1,5 +1,6 @@
 package com.plivo.api.models.conference;
 
+import com.plivo.api.PlivoClient;
 import okhttp3.ResponseBody;
 import retrofit2.Call;
 
@@ -14,4 +15,11 @@ public class ConferenceMemberSpeakDeleter extends ConferenceMemberDeleterAction<
     return client().getApiService()
       .conferenceMemberSpeakDelete(client().getAuthId(), conferenceName, id);
   }
+
+  @Override
+  public ConferenceMemberSpeakDeleter client(final PlivoClient plivoClient) {
+    this.plivoClient = plivoClient;
+    return this;
+  }
+
 }