+++ src/test/java/com/plivo/api/AccountTest.java
@@ -25,6 +25,17 @@ public class AccountTest extends BaseTest {
   }
 
   @Test
+  public void accountGetWithClientShouldSucceed() throws Exception {
+    String response = expectResponse("accountGetResponse.json", 200);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    Account account = Account.getter().client(client)
+      .get();
+    assertEquals(account.getId(), account.getAuthId());
+
+    assertRequest("GET", "");
+  }
+
+  @Test
   public void accountModifyShouldSucceed() throws Exception {
     String fixtureName = "accountModifyResponse.json";
 
@@ -38,6 +49,27 @@ public class AccountTest extends BaseTest {
       .name("test")
       .city("test")
       .update();
+
+    RecordedRequest recordedRequest = server.takeRequest();
+    assertEquals("POST", recordedRequest.getMethod());
+    assertEquals(String.format("/Account/%s/", authId),
+      recordedRequest.getPath());
+  }
+
+  @Test
+  public void accountModifyWithClientShouldSucceed() throws Exception {
+    String fixtureName = "accountModifyResponse.json";
+
+    server.enqueue(new MockResponse()
+      .setResponseCode(202)
+      .setBody(loadFixture(fixtureName))
+    );
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    Account.updater()
+      .address("test")
+      .name("test")
+      .city("test")
+      .client(client).update();
 
     RecordedRequest recordedRequest = server.takeRequest();
     assertEquals("POST", recordedRequest.getMethod());
@@ -65,12 +97,38 @@ public class AccountTest extends BaseTest {
   }
 
   @Test
+  public void subaccountCreateWithClientShouldSucceed() throws Exception {
+
+    expectResponse("subaccountCreateResponse.json", 200);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    Subaccount.creator("Test")
+      .enabled(true)
+      .client(client).create();
+
+    assertRequest("POST", "Subaccount/");
+  }
+
+  @Test
   public void subaccountGetShouldSucceed() throws Exception {
     String fixtureName = "subaccountGetResponse.json";
     String subauthId = "SAODNKNDDMY2EXY2JKMG";
     expectResponse(fixtureName, 200);
 
     Subaccount subaccount = Subaccount.getter(subauthId).get();
+
+    assertEquals(subaccount.getId(), subaccount.getAuthId());
+
+    assertRequest("GET", "Subaccount/%s/", subauthId);
+    assertEquals(subauthId, subaccount.getAuthId());
+  }
+
+  @Test
+  public void subaccountGetWithClientShouldSucceed() throws Exception {
+    String fixtureName = "subaccountGetResponse.json";
+    String subauthId = "SAODNKNDDMY2EXY2JKMG";
+    expectResponse(fixtureName, 200);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    Subaccount subaccount = Subaccount.getter(subauthId).client(client).get();
 
     assertEquals(subaccount.getId(), subaccount.getAuthId());
 
@@ -91,6 +149,19 @@ public class AccountTest extends BaseTest {
   }
 
   @Test
+  public void subaccountListWithClientShouldSucceed() throws Exception {
+    String fixtureName = "subaccountListResponse.json";
+
+    expectResponse(fixtureName, 200);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Subaccount.lister().client(client)
+      .list();
+
+    assertRequest("GET", "Subaccount/");
+  }
+
+  @Test
   public void subaccountModifyShouldSucceed() throws Exception {
     String fixtureName = "subaccountModifyResponse.json";
     String subauthId = "SAMTVIYJDIYWYYMZHLYZ";
@@ -105,12 +176,40 @@ public class AccountTest extends BaseTest {
   }
 
   @Test
+  public void subaccountModifyWithClientShouldSucceed() throws Exception {
+    String fixtureName = "subaccountModifyResponse.json";
+    String subauthId = "SAMTVIYJDIYWYYMZHLYZ";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    expectResponse(fixtureName, 202);
+
+    Subaccount.updater(subauthId, "name")
+      .enabled(true)
+      .client(client)
+      .update();
+
+    assertRequest("POST", "Subaccount/%s/", subauthId);
+  }
+
+  @Test
   public void subaccountDeleteShouldSucceed() throws Exception {
     expectResponse(null, 204);
 
     String subauthId = "SAMTVIYJDIYWYYMZHLYZ";
 
     Subaccount.deleter(subauthId).delete();
+
+    assertRequest("DELETE", "Subaccount/%s/", subauthId);
+  }
+
+  @Test
+  public void subaccountDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse(null, 204);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    String subauthId = "SAMTVIYJDIYWYYMZHLYZ";
+
+    Subaccount.deleter(subauthId).client(client).delete();
 
     assertRequest("DELETE", "Subaccount/%s/", subauthId);
   }