+++ src/test/java/com/plivo/api/ConferenceTest.java
@@ -18,11 +18,38 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceListWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceListResponse.json", 200);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    Conference.listGetter()
+      .client(client)
+      .get();
+
+    assertRequest("GET", "Conference/");
+  }
+
+  @Test
   public void conferenceGetShouldSucceed() throws Exception {
     expectResponse("conferenceGetResponse.json", 200);
     final String conferenceId = "conferenceId";
 
     Conference conference = Conference.getter(conferenceId)
+      .get();
+
+
+    assertEquals(conference.getId(), conference.getConferenceName());
+
+    assertRequest("GET", "Conference/%s/", conferenceId);
+  }
+
+  @Test
+  public void conferenceGetWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceGetResponse.json", 200);
+    final String conferenceId = "conferenceId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference conference = Conference.getter(conferenceId)
+      .client(client)
       .get();
 
 
@@ -43,11 +70,36 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceDeleteResponse.json", 204);
+    final String conferenceId = "conferenceId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.deleter(conferenceId)
+      .client(client)
+      .delete();
+
+    assertRequest("DELETE", "Conference/%s/", conferenceId);
+  }
+
+  @Test
   public void conferenceDeleteAllShouldSucceed() throws Exception {
     expectResponse("conferenceDeleteAllResponse.json", 204);
     final String conferenceId = "conferenceId";
 
     Conference.allDeleter()
+      .delete();
+
+    assertRequest("DELETE", "Conference/");
+  }
+
+  @Test
+  public void conferenceDeleteAllWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceDeleteAllResponse.json", 204);
+    final String conferenceId = "conferenceId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.allDeleter().client(client)
       .delete();
 
     assertRequest("DELETE", "Conference/");
@@ -66,12 +118,39 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceMemberDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberDeleteResponse.json", 204);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberHangupper(confId, memberId)
+      .client(client)
+      .hangup();
+
+    assertRequest("DELETE", "Conference/%s/Member/%s/", confId, memberId);
+  }
+
+  @Test
   public void conferenceMemberKickCreateShouldSucceed() throws Exception {
     expectResponse("conferenceMemberKickCreateResponse.json", 202);
     final String confId = "confId";
     final String memberId = "memberId";
 
     Conference.memberKicker(confId, memberId)
+      .kick();
+
+    assertRequest("POST", "Conference/%s/Member/%s/Kick/", confId, memberId);
+  }
+
+  @Test
+  public void conferenceMemberKickCreateWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberKickCreateResponse.json", 202);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberKicker(confId, memberId).client(client)
       .kick();
 
     assertRequest("POST", "Conference/%s/Member/%s/Kick/", confId, memberId);
@@ -90,14 +169,38 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceMemberSpeakCreateWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberSpeakCreateResponse.json", 202);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberSpeaker(confId, memberId, "test").client(client)
+      .speak();
+
+    assertRequest("POST", "Conference/%s/Member/%s/Speak/", confId, memberId);
+  }
+
+  @Test
   public void conferenceMemberPlayCreateShouldSucceed() throws Exception {
     expectResponse("conferenceMemberPlayCreateResponse.json", 202);
     final String confId = "confId";
     final String memberId = "memberId";
-
     Conference.memberPlayer(confId, memberId, "http://www.example.com/")
       .play();
+    assertRequest("POST", "Conference/%s/Member/%s/Play/", confId, memberId);
+  }
 
+  @Test
+  public void conferenceMemberPlayCreateWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberPlayCreateResponse.json", 202);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberPlayer(confId, memberId, "http://www.example.com/")
+      .client(client)
+      .play();
     assertRequest("POST", "Conference/%s/Member/%s/Play/", confId, memberId);
   }
 
@@ -108,6 +211,20 @@ public class ConferenceTest extends BaseTest {
     final String memberId = "memberId";
 
     Conference.memberMuter(confId, memberId)
+      .mute();
+
+    assertRequest("POST", "Conference/%s/Member/%s/Mute/", confId, memberId);
+  }
+
+  @Test
+  public void conferenceMemberMuteCreateWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberMuteCreateResponse.json", 202);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberMuter(confId, memberId)
+      .client(client)
       .mute();
 
     assertRequest("POST", "Conference/%s/Member/%s/Mute/", confId, memberId);
@@ -125,6 +242,18 @@ public class ConferenceTest extends BaseTest {
     assertRequest("POST", "Conference/%s/Member/%s/Deaf/", confId, memberId);
   }
 
+  @Test
+  public void conferenceMemberDeafWithClientCreateShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberDeafCreateResponse.json", 202);
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    final String confId = "confId";
+    final String memberId = "memberId";
+
+    Conference.memberDeafer(confId, memberId).client(client)
+      .deaf();
+
+    assertRequest("POST", "Conference/%s/Member/%s/Deaf/", confId, memberId);
+  }
 
   @Test
   public void conferenceMemberSpeakDeleteShouldSucceed() throws Exception {
@@ -133,6 +262,20 @@ public class ConferenceTest extends BaseTest {
     final String memberId = "memberId";
 
     Conference.memberSpeakStopper(confId, memberId)
+      .stop();
+
+    assertRequest("DELETE", "Conference/%s/Member/%s/Speak/", confId, memberId);
+  }
+
+  @Test
+  public void conferenceMemberSpeakDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberSpeakDeleteResponse.json", 204);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberSpeakStopper(confId, memberId)
+      .client(client)
       .stop();
 
     assertRequest("DELETE", "Conference/%s/Member/%s/Speak/", confId, memberId);
@@ -151,12 +294,40 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceMemberPlayDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberPlayDeleteResponse.json", 204);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberPlayStopper(confId, memberId)
+      .client(client)
+      .stop();
+
+    assertRequest("DELETE", "Conference/%s/Member/%s/Play/", confId, memberId);
+  }
+
+  @Test
   public void conferenceMemberMuteDeleteShouldSucceed() throws Exception {
     expectResponse("conferenceMemberMuteDeleteResponse.json", 204);
     final String confId = "confId";
     final String memberId = "memberId";
 
     Conference.memberMuteStopper(confId, memberId)
+      .stop();
+
+    assertRequest("DELETE", "Conference/%s/Member/%s/Mute/", confId, memberId);
+  }
+
+  @Test
+  public void conferenceMemberMuteDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberMuteDeleteResponse.json", 204);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.memberMuteStopper(confId, memberId)
+      .client(client)
       .stop();
 
     assertRequest("DELETE", "Conference/%s/Member/%s/Mute/", confId, memberId);
@@ -175,6 +346,18 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceMemberDeafDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceMemberDeafDeleteResponse.json", 204);
+    final String confId = "confId";
+    final String memberId = "memberId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+    Conference.memberDeafStopper(confId, memberId).client(client)
+      .stop();
+
+    assertRequest("DELETE", "Conference/%s/Member/%s/Deaf/", confId, memberId);
+  }
+
+  @Test
   public void conferenceRecordCreateShouldSucceed() throws Exception {
     expectResponse("conferenceRecordCreateResponse.json", 202);
     final String confId = "confId";
@@ -186,11 +369,37 @@ public class ConferenceTest extends BaseTest {
   }
 
   @Test
+  public void conferenceRecordCreateWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceRecordCreateResponse.json", 202);
+    final String confId = "confId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.recorder(confId)
+      .client(client)
+      .record();
+
+    assertRequest("POST", "Conference/%s/Record/", confId);
+  }
+
+  @Test
   public void conferenceRecordDeleteShouldSucceed() throws Exception {
     expectResponse("conferenceRecordDeleteResponse.json", 204);
     final String confId = "confId";
 
     Conference.recordStopper(confId)
+      .stop();
+
+    assertRequest("DELETE", "Conference/%s/Record/", confId);
+  }
+
+  @Test
+  public void conferenceRecordDeleteWithClientShouldSucceed() throws Exception {
+    expectResponse("conferenceRecordDeleteResponse.json", 204);
+    final String confId = "confId";
+    PlivoClient client = new PlivoClient("MA123456789012345678", "Zmxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+
+    Conference.recordStopper(confId)
+      .client(client)
       .stop();
 
     assertRequest("DELETE", "Conference/%s/Record/", confId);